chore : merge

+++ server/src/main/java/seb39_40/coffeewithme/tag/service/TagService.java
@@ -3,6 +3,8 @@ package seb39_40.coffeewithme.tag.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import seb39_40.coffeewithme.cafe.domain.Cafe;
+import seb39_40.coffeewithme.cafe.domain.CafeTag;
 import seb39_40.coffeewithme.exception.BusinessLogicException;
 import seb39_40.coffeewithme.exception.ExceptionCode;
 import seb39_40.coffeewithme.review.domain.Review;
@@ -11,8 +13,7 @@ import seb39_40.coffeewithme.tag.domain.Tag;
 import seb39_40.coffeewithme.tag.repository.ReviewTagRepository;
 import seb39_40.coffeewithme.tag.repository.TagRepository;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
@@ -29,8 +30,9 @@ public class TagService {
         return tagRepository.findByName(name).orElseThrow(() -> new BusinessLogicException(ExceptionCode.TAG_NOT_FOUND));
     }
 
+    @Transactional
     public List<ReviewTag> createReviewTag(Review review, List<String> tags) {
-        List<ReviewTag> reviewTags = tags.stream().map(tag -> {
+        List<ReviewTag> reviewTags = tags.stream().distinct().map(tag -> {
             Tag t;
             Optional<Tag> tempTag = tagRepository.findByName(tag);
             if (tempTag.isEmpty()){
@@ -50,19 +52,55 @@ public class TagService {
         return reviewTags;
     }
 
+    @Transactional
+    public List<CafeTag> createCafeTag(Cafe cafe, List<String> tags){
+        List<CafeTag> cafeTags = tags.stream().distinct().map(tag -> {
+            Tag t;
+            Optional<Tag> tempTag = tagRepository.findByName(tag);
+            if (tempTag.isEmpty()){
+                t = new Tag();
+                t.setName(tag);
+                t.setCategory(Tag.Category.NONE);
+                tagRepository.save(t);
+                t = tagRepository.findByName(tag).get();
+            }
+            else t = tempTag.get();
+
+            CafeTag cafeTag = new CafeTag();
+            cafeTag.setCafe(cafe);
+            cafeTag.setTag(t);
+            return cafeTag;
+        }).collect(Collectors.toList());
+        return cafeTags;
+    }
+
     public List<ReviewTag> findByReviewId(Long id) {
         return reviewTagRepository.findByReviewId(id);
     }
 
     @Transactional
-    public void deleteReviewTag(Review review){
+    public void updateReviewTag(Review review, List<String> tags){
        List<ReviewTag> reviewTags = reviewTagRepository.findByReviewId(review.getId());
-        System.out.println(reviewTags.size());
+
+       Set<String> tagSet = new HashSet<>(tags); // 빠른 조회를 위해 Set 변환
+       Set<String> existTag = new HashSet<>(); // 수정되지 않은(유지) 태그 보관을 위한 List
        for (ReviewTag reviewTag : reviewTags){
-           System.out.println("삭제");
-           System.out.println(reviewTag.getTag().getCategory());
+           if (tagSet.contains(reviewTag.getTag().getName())) {
+               existTag.add(reviewTag.getTag().getName());
+               continue;
+           }
+
            reviewTagRepository.delete(reviewTag);
        }
+
+       List<String> newTags = new ArrayList<>();
+       for (String tag: tags) {
+           if(existTag.contains(tag)) continue;
+           newTags.add(tag);
+        }
+
+        List<ReviewTag> result = createReviewTag(review, newTags);
+        review.setReviewTags(result);
     }
